requirements.txt
pandas>=2.1
numpy>=1.26
matplotlib>=3.8
nltk>=3.9
pyyaml>=6.0